requirements.txt
discord.py[voice] == 2.5.2